requirements.txt
opencv-python==4.8.1.78
numpy>=1.21.0,<2.0
pyyaml>=6.0
ultralytics>=8.2.0
deep-sort-realtime>=1.3.2
easyocr>=1.7.0
pyttsx3>=2.90
transformers>=4.30.0
torch>=1.13.0
pillow>=9.0.0
